# Restructures parsing logic for performance