… route-geometry-sweden-west-coast.polyline6