## Intro 

This change removes the translation to `Vec<char>` inside of the loop which makes a massive difference in decoding performance.

Further, I modularized the code, similar to a [python-decoding library does it](https://github.com/frederickjansen/polyline/blob/77893079ea38bfea3dc03135f94e2a27efffafbd/polyline/codec.py#L27-L52)

I added a new criterion benchmark which shows a 78% performance increase for the given small polyline,
and several orders of magnitude for a larger polyline.

## Benchmark of decoding of a small polyline 
```
bench polyline6 decoding
                        time:   [33.483 ns 33.527 ns 33.568 ns]
                        change: [-78.983% -78.787% -78.592%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) low mild
```
## Benchmark of decoding of a large polyline 
```
bench HUGE polyline6 decoding
                        time:   [20.007 us 20.012 us 20.018 us]
                        change: [-99.975% -99.974% -99.974%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
```

The large polyline decoding went from 78milliseconds to 20 _micro-seconds_!