@urschrei 2023-03-11
Hi @purew, sorry I didn't see this until now! We have two open decoding perf-related PRs, and there doesn't seem to be much overlap, but we'll figure out a way forward ASAP.