@urschrei 2023-03-12
This PR has a 30 % improvement over https://github.com/georust/polyline/pull/26, so going to merge this and release.